<refactoring>Change file name posts**. => post**
