+++ back/main_sever/src/posts/const/default-post-find-options.const.ts
@@ -1,5 +1,5 @@
 import { FindManyOptions } from 'typeorm';
-import { PostModel } from '../entities/posts.entity';
+import { PostModel } from '../entities/post.entity';
 
 // where 옵션은 함수안에서 사용
 export const DEFAULT_POST_FIND_OPTIONS: FindManyOptions<PostModel> = {
