+++ back/main_sever/src/posts/dto/create-post.dto.ts
@@ -1,7 +1,7 @@
 import { IsArray, IsOptional, IsString } from 'class-validator';
 import { PickType } from '@nestjs/mapped-types';
 
-import { PostModel } from '../entities/posts.entity';
+import { PostModel } from '../entities/post.entity';
 
 // Pick, Omit, Partial -> Type 반환
 // PickType, OmitType, ParitalType -> 값을 반환
