+++ back/main_sever/src/posts/dto/update-post.dto.ts
@@ -1,5 +1,5 @@
 import { PartialType, PickType } from "@nestjs/mapped-types";
-import { PostModel } from "../entities/posts.entity";
+import { PostModel } from "../entities/post.entity";
 import { CreatePostDto } from "./create-post.dto";
 import { IsOptional, IsString } from "class-validator";
 import { stringValidationMessage } from "src/common/validation-message/string-validation.message";
