+++ back/main_sever/src/posts/post.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { PostsService } from './posts.service';
+import { PostsService } from './post.service';
 
 describe('PostsService', () => {
   let service: PostsService;
